feat(inheritance): fix PR review changes
