+++ packages/ui/src/components/molecules/index.ts
@@ -46,4 +46,4 @@ export * from './Accordion';
 export * from './SimpleJsonView';
 export * from './OTPInput';
 export * from './OneInMany';
-export * from './Coupon';
+export * from './CouponInput';
